Integrate the xvfb_it decorators into the testing module and use everywhere.
