+++ dipy/viz/tests/test_fvtk_utils.py
