+++ dipy/viz/tests/test_fvtk_window.py
@@ -1,18 +1,7 @@
 import numpy as np
 from dipy.viz import actor, window
 import numpy.testing as npt
-
-
-def xvfb_it(my_test):
-    def run_with_xvfb():
-        if is_travis:
-            from xvfbwrapper import Xvfb
-            display = Xvfb()
-            display.start()
-        my_test()
-        if is_travis:
-            display.stop()
-    return run_with_xvfb
+from dipy.testing.decorators import xvfb_it
 
 
 @xvfb_it
